Refactor: Tasks로부터 TaskItem 분리
