+++ src/components/TaskItem/index.js
@@ -1,0 +1,23 @@
+/* eslint-disable react/prop-types */
+import classNames from 'classnames'
+import styles from './TaskItem.module.scss'
+
+function TaskItem({ isDone, category, content, onClick }) {
+  return (
+    <li>
+      <button
+        type='button'
+        aria-label='Toggle Task'
+        className={classNames(
+          styles.doneCircle,
+          category === 'personal' ? styles.personal : styles.business,
+          isDone ? styles.done : ''
+        )}
+        onClick={onClick}
+      />
+      <div className={classNames(isDone ? styles.done : '')}>{content}</div>
+    </li>
+  )
+}
+
+export default TaskItem
