+++ src/components/Tasks/index.js
@@ -1,8 +1,7 @@
 /* eslint-disable react/prop-types */
-import classNames from 'classnames'
-
 import SectionTitle from '../SectionTitle'
 import styles from './Tasks.module.scss'
+import TaskItem from '../TaskItem'
 
 function Tasks({ tasks, setTasks }) {
   const handleToggleTask = (id) => {
@@ -28,21 +27,3 @@ function Tasks({ tasks, setTasks }) {
 }
 
 export default Tasks
-
-function TaskItem({ isDone, category, content, onClick }) {
-  return (
-    <li>
-      <button
-        type='button'
-        aria-label='Toggle Task'
-        className={classNames(
-          styles.doneCircle,
-          category === 'personal' ? styles.personal : styles.business,
-          isDone ? styles.done : ''
-        )}
-        onClick={onClick}
-      />
-      <div className={classNames(isDone ? styles.done : '')}>{content}</div>
-    </li>
-  )
-}
